+++ lib/widgets/numeric_fields/card_settings_double.dart
@@ -38,9 +38,6 @@ class CardSettingsDouble extends CardSettingsText {
     ValueChanged<String> onFieldSubmitted,
     List<TextInputFormatter> inputFormatters,
     bool showMaterialonIOS: false,
-
-    ///Shows a [red] [Border] around the [CupertinoTextField] since the [validator] does not exist
-    bool showErrorIOS: false,
   }) : super(
           key: key,
           label: label,
@@ -57,7 +54,6 @@ class CardSettingsDouble extends CardSettingsText {
           enabled: enabled,
           autofocus: autofocus,
           obscureText: obscureText,
-          showErrorIOS: showErrorIOS,
           autocorrect: autocorrect,
           autovalidate: autovalidate,
           validator: (value) {