+++ lib/widgets/text_fields/card_settings_password.dart
@@ -39,9 +39,6 @@ class CardSettingsPassword extends CardSettingsText
     ValueChanged<String> onFieldSubmitted,
     List<TextInputFormatter> inputFormatters,
     bool showMaterialonIOS: false,
-
-    ///Shows a [red] [Border] around the [CupertinoTextField] since the [validator] does not exist
-    bool showErrorIOS: false,
   }) : super(
           key: key,
           label: label,
@@ -61,7 +58,6 @@ class CardSettingsPassword extends CardSettingsText
           obscureText: obscureText,
           autocorrect: autocorrect,
           autovalidate: autovalidate,
-          showErrorIOS: showErrorIOS,
           validator: validator,
           onSaved: onSaved,
           onChanged: onChanged,